tech(mon-pix): remove call to GET /assessment/:id/next
