+++ mon-pix/app/routes/assessments/resume.js
@@ -26,13 +26,10 @@ export default class ResumeRoute extends Route {
     if (!transition.to.parent.params.assessment_id) {
       this.assessmentHasNoMoreQuestions = false;
     } else {
-      const assessment = this.hasSeenCheckpoint
-        ? await this.store.findRecord('assessment', transition.to.parent.params.assessment_id, {
-            backgroundReload: false,
-          })
-        : await this.store.queryRecord('challenge', {
-            assessmentId: transition.to.parent.params.assessment_id,
-          });
+      const assessment = await this.store.findRecord('assessment', transition.to.parent.params.assessment_id, {
+        backgroundReload: false,
+        reload: !this.hasSeenCheckpoint,
+      });
       this.assessmentHasNoMoreQuestions = !assessment.nextChallenge;
     }
   }
